requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
python-dateutil==2.8.2
pymysql==1.1.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
supabase==2.0.3